removed indirection and made it possible to set the debug level on the purchases object directly
