+++ Purchases/Public/RCLogger.h
@@ -7,15 +7,9 @@
 //
 
 #import <Foundation/Foundation.h>
+#import "RCLogLevel.h"
 
 NS_ASSUME_NONNULL_BEGIN
-
-typedef NS_ENUM(NSInteger, RCLogLevel) {
-    RCLogLevelDebug = 0,
-    RCLogLevelInfo = 1,
-    RCLogLevelWarn = 2,
-    RCLogLevelError = 3,
-} NS_SWIFT_NAME(Purchases.LogLevel);
 
 NS_SWIFT_NAME(Purchases.Logger)
 @interface RCLogger : NSObject
